refactor(usage): neutralize credit usage service filename
