+++ turbo/apps/api/src/signals/services/agent-webhook-complete.service.ts
@@ -36,7 +36,7 @@ import {
 import { lockChatQueueThread } from "./chat-event-queue.service";
 import { projectLegacyCheckpointStorage } from "./storage-legacy-projection.service";
 import { maybeEmitRunUsageEvent$ } from "./chat-usage-event.service";
-import { processOrgUsageEvents$ } from "./zero-credit-usage.service";
+import { processOrgUsageEvents$ } from "./credit-usage.service";
 import { drainOrgQueue$ } from "./zero-run-queue.service";
 
 type WebhookCompleteBody = z.infer<
