+++ turbo/apps/api/src/signals/services/managed-usage.service.ts
@@ -19,7 +19,7 @@ import {
 import { writeDb$ } from "../external/db";
 import { resolveUsageAllowanceAvailability } from "./usage-allowance.service";
 import { getSpendableUsagePackCredits } from "./usage-pack-credit.service";
-import { processOrgUsageEvents$ } from "./zero-credit-usage.service";
+import { processOrgUsageEvents$ } from "./credit-usage.service";
 
 export interface ManagedUsageErrorResponse {
   readonly status: 402 | 503;
